# Avoid static overhead for switch-over-string optimization