--- a/src/Compilers/Core/Portable/Diagnostic/ProgrammaticSuppressionInfo.cs
+++ b/src/Compilers/Core/Portable/Diagnostic/ProgrammaticSuppressionInfo.cs
@@ -27,7 +27,7 @@ namespace Microsoft.CodeAnalysis.Diagnostics
             }
 
             return other != null &&
-                this.Suppressions.SetEquals(other.Suppressions);
+                this.Suppressions.SetEqualsWithoutIntermediateHashSet(other.Suppressions);
         }
 
         public override bool Equals(object? obj)