* Avoid retaining large, sparse arrays
* Avoid allocating intermediate hash sets (https://github.com/dotnet/runtime/issues/90986)

Fixes [AB#1874059](https://devdiv.visualstudio.com/0bdbc590-a062-4c3f-b0f6-9383f67865ee/_workitems/edit/1874059)

### Baseline

```
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.22621
Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
.NET SDK=8.0.100-preview.7.23376.3
  [Host]     : .NET 7.0.10 (7.0.1023.36312), X64 RyuJIT
  Job-SDXWZP : .NET 7.0.10 (7.0.1023.36312), X64 RyuJIT

InvocationCount=1  RunStrategy=Throughput  UnrollFactor=1
```

|   Method | SwitchCount |         Mean |      Error |     StdDev |        Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|--------- |------------ |-------------:|-----------:|-----------:|-------------:|-----------:|----------:|----------:|
| BindFile |         100 |     21.11 ms |   0.407 ms |   0.380 ms |            - |          - |         - |      3 MB |
| BindFile |        1000 |    472.88 ms |   9.296 ms |   9.130 ms |   21000.0000 |  8000.0000 | 1000.0000 |    125 MB |
| BindFile |        5000 | 10,243.52 ms |  52.573 ms |  49.177 ms |  292000.0000 | 45000.0000 | 4000.0000 |  2,746 MB |
| BindFile |       10000 | 45,215.70 ms | 583.825 ms | 546.110 ms | 1042000.0000 | 74000.0000 | 5000.0000 | 11,298 MB |

### Avoid retaining large arrays with few elements

```
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.22621
Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
.NET SDK=8.0.100-preview.7.23376.3
  [Host]     : .NET 7.0.10 (7.0.1023.36312), X64 RyuJIT
  Job-CLGAKY : .NET 7.0.10 (7.0.1023.36312), X64 RyuJIT

InvocationCount=1  RunStrategy=Throughput  UnrollFactor=1
```

|   Method | SwitchCount |         Mean |      Error |     StdDev |        Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|--------- |------------ |-------------:|-----------:|-----------:|-------------:|-----------:|----------:|----------:|
| BindFile |         100 |     21.07 ms |   0.416 ms |   0.541 ms |            - |          - |         - |      3 MB |
| BindFile |        1000 |    456.06 ms |   8.680 ms |   8.525 ms |   21000.0000 |  5000.0000 | 1000.0000 |    125 MB |
| BindFile |        5000 | 10,411.66 ms |  80.092 ms |  74.918 ms |  296000.0000 | 40000.0000 | 8000.0000 |  2,746 MB |
| BindFile |       10000 | 42,928.17 ms | 326.267 ms | 254.728 ms | 1046000.0000 | 72000.0000 | 9000.0000 | 11,298 MB |

### Avoid intermediate HashSet<T>

```
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.22621
Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
.NET SDK=8.0.100-preview.7.23376.3
  [Host]     : .NET 7.0.10 (7.0.1023.36312), X64 RyuJIT
  Job-ZMBVHX : .NET 7.0.10 (7.0.1023.36312), X64 RyuJIT

InvocationCount=1  RunStrategy=Throughput  UnrollFactor=1
```

|   Method | SwitchCount |         Mean |      Error |     StdDev |       Median |       Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|--------- |------------ |-------------:|-----------:|-----------:|-------------:|------------:|-----------:|----------:|----------:|
| BindFile |         100 |     20.68 ms |   0.413 ms |   0.843 ms |     20.30 ms |           - |          - |         - |      3 MB |
| BindFile |        1000 |    417.38 ms |   7.990 ms |  10.936 ms |    418.08 ms |  20000.0000 |  5000.0000 | 1000.0000 |    114 MB |
| BindFile |        5000 |  8,663.36 ms |  89.757 ms |  79.568 ms |  8,660.18 ms | 253000.0000 | 41000.0000 | 7000.0000 |  2,484 MB |
| BindFile |       10000 | 37,244.51 ms | 423.100 ms | 375.067 ms | 37,137.94 ms | 978000.0000 | 72000.0000 | 9000.0000 | 10,253 MB |