@333fred 2023-08-23
@sharwell were you able to collect the full benchmark numbers from this change? Looking at the baseline->commit 1 diff in your numbers, I'm unsure why this is an overall win.